Reducing ModalInscripcionClase
