+++ FrontEnd/src/Pages/AdministratorClassRegister/ShowClass.jsx
@@ -1,12 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import {
-  Button,
-  Modal,
-  TextField,
-  Typography,
-} from '@mui/material';
-import MenuItem from '@mui/material/MenuItem';
-import WaitList from '../../Components/Clase/WaitList';
 import { getWaitList } from '../../api/waitList';
 import { getStudents } from '../../api/students';
 import { getPeriodos } from '../../api/Periodos';
@@ -15,14 +7,14 @@ import {
   createClass, deleteClasses, getClasses, updateClass,
 } from '../../api/classes.js';
 import {
-  classAtributes, dayAtributes, niveloptions, classTemplate, nivelesMapa, claseActualDefault, profesorVacioInscripcion,
+  classTemplate, nivelesMapa, claseActualDefault, profesorVacioInscripcion,
 } from '../../utils/constants';
 import HeaderInscripcionClase from '../../Components/Clase/HeaderInscripcionClase';
 import { mapNiveles } from '../../utils/utilFunctions';
 import BodyInscripcionClase from '../../Components/Clase/BodyInscripcionClase';
+import ModalInscripcionClase from '../../Components/Clase/ModalInscripcionClase';
 
 export default function ShowClass() {
-  let id = '';
   const [dataPeriodo, setDataPeriodo] = useState([]);
   const [data, setData] = useState([]);
   const [profesorList, setProfesorList] = useState([profesorVacioInscripcion]);
@@ -162,7 +154,6 @@ export default function ShowClass() {
     if (caso === 'Crear') {
       setCurrentProfesor(profesorVacioInscripcion);
     } else if (caso === 'Editar') {
-      id = consola._id;
       setClaseActual(consola);
       profesorList.forEach((e) => {
         if (e.nombreCompleto === consola.nombreCompleto) {
@@ -171,7 +162,6 @@ export default function ShowClass() {
       });
     } else if (caso === 'Eliminar') {
       setClaseActual(consola);
-      id = consola._id;
     }
     setCurrentOperation(caso);
     console.log(caso);
@@ -185,270 +175,39 @@ export default function ShowClass() {
     setOpenModal(!openModal);
   };
 
-  const getClassWaitList = (clase) => {
-    let waitList = [];
-    let students = [];
-    const result = [];
-    getStudents().then((response) => response.json()).then((data) => {
-      students = data;
-    }).then(() => {
-      getWaitList().then((response) => response.json()).then((data) => {
-        waitList = data.filter((lista) => lista.idClase === clase._id);
-        waitList.map((inWaitList) => {
-          for (let i = 0; i < students.length; i++) {
-            if (inWaitList.idAlumno === students[i]._id) {
-              result.push({
-                _id: inWaitList._id,
-                studentName: `${students[i].nombre} ${students[i].apellido_paterno} ${students[i].apellido_materno}`,
-                time_stamp: inWaitList.time_stamp,
-              });
-            }
-          }
-        });
-        result.sort((a, b) => (a > b ? 1 : a < b ? -1 : 0));
-        setCurrentWaitList(result);
-        setCurrentClase(clase);
-        seleccionarClase(clase, 'AbrirWaitList');
+  const getClassWaitList = async (clase) => {
+    try {
+      const studentsList = await getStudents();
+      const students = await studentsList.json();
+      const studentsById = students.reduce((obj, student) => {
+        obj[student._id] = student;
+        return obj;
+      }, {});
+  
+      const responseWaitList = await getWaitList();
+      const waitList = await responseWaitList.json();
+      const result = [];
+  
+      waitList.forEach((inWaitList) => {
+        if (inWaitList.idClase === clase._id && studentsById[inWaitList.idAlumno]) {
+          const student = studentsById[inWaitList.idAlumno];
+          result.push({
+            _id: inWaitList._id,
+            studentName: `${student.nombre} ${student.apellido_paterno} ${student.apellido_materno}`,
+            time_stamp: inWaitList.time_stamp,
+          });
+        }
       });
-    });
+  
+      result.sort((a, b) => (a > b ? 1 : a < b ? -1 : 0));
+      setCurrentWaitList(result);
+      setCurrentClase(clase);
+      seleccionarClase(clase, 'AbrirWaitList');
+    } catch (error) {
+      console.error(error);
+    }
   };
-
-  const bodyEditar = (
-    <div
-      style={{
-        position: 'absolute',
-        width: 520,
-        height: '95vh',
-        backgroundColor: '#fefefd',
-        top: '48%',
-        left: '50%',
-        transform: 'translate(-48%, -50%)',
-        border: '4px solid  rgb(165, 165, 180)',
-        margin: 'auto',
-        borderRadius: '10px',
-        padding: '20px',
-        display: 'flex',
-        justifyContent: 'center',
-        flexWrap: 'wrap',
-        overflowY: 'scroll',
-      }}
-    >
-      <h3
-        style={{
-          paddingBottom: '15px',
-          marginTop: '5px',
-          fontFamily: 'arial',
-          width: '100%',
-        }}
-        align="center"
-      >
-        Actualizar una clase
-      </h3>
-      {classAtributes.map((atribute) => (
-        <TextField
-          style={{
-            paddingBottom: '15px',
-            fontFamily: 'arial',
-            marginRight: 10,
-            width: '40%',
-          }}
-          label={atribute.value}
-          onChange={(e) => {
-            handleChange(e);
-          }}
-          name={atribute.key}
-          key={atribute.key}
-          value={clase[atribute.key]}
-          autoFocus
-        />
-      ))}
-      <TextField
-        style={{
-          paddingBottom: '15px',
-          fontFamily: 'arial',
-          marginRight: 10,
-          width: '40%',
-        }}
-        label="Modalidad"
-        value={clase.modalidad}
-        name="modalidad"
-        onChange={(e) => {
-          handleChange(e);
-        }}
-        select
-      >
-        {['presencial', 'online'].map((e) => (
-          <MenuItem value={e} key={e}>
-            {e}
-          </MenuItem>
-        ))}
-      </TextField>
-      <TextField
-        style={{
-          paddingBottom: '15px',
-          fontFamily: 'arial',
-          marginRight: 10,
-          width: '40%',
-        }}
-        label="Nivel"
-        value={clase.niveles}
-        name="niveles"
-        onChange={(e) => {
-          handleChange(e);
-        }}
-        select
-      >
-        {niveloptions.map((e) => (
-          <MenuItem value={e} key={e}>
-            {e}
-          </MenuItem>
-        ))}
-      </TextField>
-      <div
-        style={{
-          width: '100%',
-          borderTop: '1px solid gray',
-          paddingTop: '5px',
-          display: 'flex',
-          justifyContent: 'center',
-          flexWrap: 'wrap',
-        }}
-      >
-        <Typography
-          sx={{ textAlign: 'center', marginTop: '10px', width: '100%' }}
-        >
-          Horarios
-        </Typography>
-        {dayAtributes.map((atribute) => (
-          <TextField
-            style={{
-              paddingBottom: '15px',
-              fontFamily: 'arial',
-              marginRight: 10,
-            }}
-            label={atribute.value}
-            onChange={(e) => {
-              handleChange(e);
-            }}
-            name={atribute.key}
-            key={atribute.key}
-            value={clase[atribute.key]}
-            autoFocus
-          />
-        ))}
-      </div>
-      <div
-        style={{
-          width: '100%',
-          borderTop: '1px solid gray',
-          paddingTop: '5px',
-        }}
-      >
-        <Typography sx={{ textAlign: 'center', marginTop: '10px' }}>
-          Datos del profesor
-        </Typography>
-        <br />
-        <TextField
-          style={{
-            paddingBottom: '15px',
-            fontFamily: 'arial',
-            marginRight: 10,
-            width: '100%',
-          }}
-          label="Profesor"
-          value={currentProfesor.nombreCompleto}
-          name="nombreCompleto"
-          onChange={(e) => {
-            handleChangeProfesor(e);
-          }}
-          select
-        >
-          {profesorList.map((e) => (
-            <MenuItem value={`${e.nombre} ${e.apellidos}`} key={e._id}>
-              {`${e.nombre} ${e.apellidos}`}
-            </MenuItem>
-          ))}
-        </TextField>
-        <TextField
-          style={{
-            paddingBottom: '15px',
-            fontFamily: 'arial',
-            marginRight: 10,
-            width: '40%',
-          }}
-          InputProps={{
-            readOnly: true,
-          }}
-          value={currentProfesor.matricula}
-          defaultValue={currentProfesor.matricula}
-          variant="filled"
-          label="matricula"
-        />
-        <TextField
-          style={{
-            paddingBottom: '15px',
-            fontFamily: 'arial',
-            marginRight: 10,
-            width: '40%',
-          }}
-          InputProps={{
-            readOnly: true,
-          }}
-          value={currentProfesor.correo}
-          defaultValue={currentProfesor.correo}
-          variant="filled"
-          label="correo"
-        />
-      </div>
-      <div align="center" style={{ width: '100%' }}>
-        <Button color="primary" onClick={modalSubmit}>
-          Editar
-        </Button>
-        <Button onClick={abrirCerrarModal} color="error">
-          Cancelar
-        </Button>
-      </div>
-    </div>
-  );
-
-  const bodyEliminar = (
-    <div
-      style={{
-        position: 'absolute',
-        width: 260,
-        height: 220,
-        backgroundColor: '#fefefd',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%, -50%)',
-        border: '4px solid  rgb(165, 165, 180)',
-        margin: 'auto',
-        borderRadius: '10px',
-        padding: '20px',
-      }}
-    >
-      <h3
-        style={{ paddingBottom: '15px', marginTop: '5px', fontFamily: 'arial' }}
-        align="center"
-      >
-        Eliminar clase
-      </h3>
-      <Typography style={{ align: 'justify', fontFamily: 'arial' }}>
-        {`Esta clase ${clase.clave} y toda su información relacionada a ella va a ser eliminada`}
-      </Typography>
-      <br />
-      <br />
-      <div align="center">
-        <Button color="error" onClick={modalSubmit}>
-          Confirmar
-        </Button>
-        <Button onClick={abrirCerrarModal} color="primary">
-          Cancelar
-        </Button>
-      </div>
-    </div>
-  );
-
+  
   return (
     <div>
       <HeaderInscripcionClase
@@ -464,11 +223,19 @@ export default function ShowClass() {
         getClassWaitList={getClassWaitList}
         seleccionarClase={seleccionarClase}
       />
-      <Modal open={openModal} onClose={abrirCerrarModal}>
-        <div>
-          {currentOperation === 'Editar' || currentOperation === 'Crear' ? bodyEditar : (currentOperation === 'AbrirWaitList' ? <WaitList clase={currentClase} waitList={currentWaitList} /> : bodyEliminar)}
-        </div>
-      </Modal>
+      <ModalInscripcionClase 
+        clase={clase}
+        currentProfesor={currentProfesor}
+        handleChange={handleChange}
+        profesorList={profesorList}
+        modalSubmit={modalSubmit}
+        currentOperation={currentOperation}
+        openModal={openModal}
+        abrirCerrarModal={abrirCerrarModal}
+        currentClase={currentClase}
+        currentWaitList={currentWaitList}
+        handleChangeProfesor={handleChangeProfesor}
+      />
     </div>
   );
 }
